// src/types/types.ts
export type dndCharSheetInputs = {
  authorId: string;
  authorEmail: string;
  characterName: string;
  gender: string;
  backgroundColor: string;
  concept: string;
  height: string;
  weight: string;
  age: string;
  race: string;
  backstory: string;
  motivation: string;
  reasonToJoin: string;
  ideal: string;
  family: string;
  playerPCs: string;
  NPCs: string;
  organisations: string;
  description: string;
  motherland: string;
  fears: string;
  habitsquirks: string;
  voice: string;
  deity: string;
  conflict: string;
  catchphrase: string;
  secret: string;
};

export type dndCharSheetInputsDb = {
  id: string;
  authorId: string;
  img: string;
  createdAt: string;
  authorEmail: string;
  characterName: string;
  gender: string;
  concept: string;
  height: string;
  age: string;
  weight: string;
  ideal: string;
  race: string;
  backstory: string;
  motivation: string;
  reasonToJoin: string;
  family: string;
  playerPCs: string;
  NPCs: string;
  organisations: string;
  description: string;
  motherland: string;
  fears: string;
  habitsquirks: string;
  voice: string;
  deity: string;
  conflict: string;
  catchphrase: string;
  secret: string;
  backgroundColor: string;
};

export type cocCharSheetInputs = {
  authorId: string;
  authorEmail: string;
  characterName: string;
  gender: string;
  concept: string;
  height: string;
  age: string;
  weight: string;
  ideal: string;
  birthday: string;
  backstory: string;
  motivation: string;
  reasonToJoin: string;
  family: string;
  playerPCs: string;
  NPCs: string;
  organisations: string;
  description: string;
  motherland: string;
  fears: string;
  habitsquirks: string;
  voice: string;
  religion: string;
  occupation: string;
  conflict: string;
  catchphrase: string;
  secret: string;
  backgroundColor: string;
  interests: string;
};

export type cocCharSheetInputsDb = {
  id: string;
  authorId: string;
  img: string;
  createdAt: string;
  authorEmail: string;
  characterName: string;
  gender: string;
  concept: string;
  height: string;
  age: string;
  weight: string;
  ideal: string;
  birthday: string;
  backstory: string;
  motivation: string;
  reasonToJoin: string;
  family: string;
  playerPCs: string;
  NPCs: string;
  organisations: string;
  description: string;
  motherland: string;
  fears: string;
  habitsquirks: string;
  voice: string;
  religion: string;
  occupation: string;
  conflict: string;
  catchphrase: string;
  secret: string;
  backgroundColor: string;
  interests: string;
};